scripts/requirements.txt
colorlog==6.6.0
tabulate==0.8.10
requests==2.28.1
google-cloud-monitoring==2.10.0
google_cloud_bigquery==2.34.3
python-gitlab==3.5.0
tqdm==4.64.1

